Moved snapshotter to snapshot package
